\name{metaPCA-package}
\alias{metaPCA-package}
\alias{metaPCA}
\docType{package}
\title{
\packageTitle{metaPCA}
}
\description{
\packageDescription{metaPCA}
}
\author{
\packageAuthor{metaPCA}

Maintainer: \packageMaintainer{metaPCA}
}
\references{
META-ANALYTIC PRINCIPAL COMPONENT ANALYSIS IN INTEGRATIVE OMICS APPLICATION
}