\name{prostate}
\docType{data}
\alias{prostate}
\title{4 prostate cancer studies}
\description{
4 prostate cancer studies comparing three classes: normal, primary, metastasis.
\tabular{lllll}{
Data Name \tab Published Year   \tab Array Platform \tab Sample Size \tab GEO Accession ID\cr
Lapointe \tab 2004   \tab cDNA \tab 112 \tab GSE3933\cr
Yu \tab 2004 \tab HG-U95Av2 \tab 108 \tab GSE6919\cr
Varambally \tab 2005 \tab HG-U133 Plus 2 \tab 19 \tab GSE3325\cr
Tomlins \tab 2007 \tab cDNA \tab 76 \tab GSE6099\cr
}
}
\usage{prostate}
\format{A list containing 4 matrices. Each matrix is gene expression data after gene filtering.}
\source{Gene Expression Omnibus (GEO)}
\references{
Lapointe, J., Li, C., Higgins, J. P., Van De Rijn, M., Bair, E., Montgomery, K., Ferrari, M., Egevad, L., Rayford, W., Bergerheim, U. et al. (2004). Gene expression profiling identifies clinically relevant subtypes of prostate cancer. Proceedings of the National Academy of Sciences of the United States of America 101 811.

Yu, Y. P., Landsittel, D., Jing, L., Nelson, J., Ren, B., Liu, L., McDonald, C., Thomas, R., Dhir, R., Finkelstein, S. et al. (2004). Gene expression alterations in prostate cancer predicting tumor aggression and preceding development of malignancy. Journal of Clinical Oncology 22 2790.

Varambally, S., Yu, J., Laxman, B., Rhodes, D. R., Mehra, R., Tomlins, S. A., Shah, R. B., Chandran, U., Monzon, F. A., Becich, M. J. et al. (2005). Integrative genomic and proteomic analysis of prostate cancer reveals signatures of metastatic progression. Cancer cell 8 393-406.

Tomlins, S. A., Mehra, R., Rhodes, D. R., Cao, X., Wang, L., Dhanasekaran, S. M., Kalyana-Sundaram, S., Wei, J. T., Rubin, M. A., Pienta, K. J. et al. (2006). Integrative molecular concept modeling of prostate cancer progression. Nature genetics 39 41-51.

}
\keyword{MetaPCA}
